+++ ExamCorrection/Entities/StudentExamPaper.cs
@@ -9,7 +9,7 @@ public class StudentExamPaper
     public string GeneratedPdfPath { get; set; } = string.Empty;
     public DateTime GeneratedAt { get; set; } = DateTime.Now;
     public float? FinalScore { get; set; }
-    public int? TotalQuestions { get; set; } = 0;          // تبدأ من 0
+    public float? TotalQuestions { get; set; } = 0;          // تبدأ من 0
     public string? QuestionDetailsJson { get; set; } = "[]"; // تبدأ كـ JSON فارغ (قائمة)
     public string? AnnotatedImageUrl { get; set; } = "";   // تبدأ كسلسلة فارغة
